remove old commands
